a few cpps and h
