+++ Logger.h
@@ -1,0 +1,21 @@
+//
+// Created by MINALP1 on 5/26/2025.
+//
+#include <iostream>
+#include <string>
+#ifndef LOGGER_H
+#define LOGGER_H
+
+
+
+class Logger {
+    // Logger - fara modificari necesare
+
+    public:
+        static void logEvent(const std::string &msg);
+
+};
+
+
+
+#endif //LOGGER_H
