+++ main.cpp
@@ -8,257 +8,15 @@
 #include <mutex>
 #include <condition_variable>
 #include <limits>
+#pragma once
+#include "Fish.h"
+#include "Logger.h"
+#include "Reward.h"
+#include "Aquarium.h"
 
-
-// Logger - fara modificari necesare
-class Logger
-{
-public:
-    static void logEvent(const std::string &msg)
-    {
-        std::cout << "[LOG] " << msg << std::endl;
-    }
-};
-
-// Clasa Fish
-class Fish
-{
-    std::string name;
-    int size;
-    int speed;
-    int growthFactor;
-
-public:
-    // Constructor cu parametri
-    Fish(const std::string &name, int size, int speed, int growthFactor = 1)
-        : name(name), size(size), speed(speed), growthFactor(growthFactor)
-    {
-    }
-
-    // Constructor de copiere
-    Fish(const Fish &f)
-        : name(f.name), size(f.size), speed(f.speed), growthFactor(f.growthFactor)
-    {
-    }
-
-    // Operator= de copiere
-    Fish &operator=(const Fish &other)
-    {
-        if (this != &other)
-        {
-            name = other.name;
-            size = other.size;
-            speed = other.speed;
-            growthFactor = other.growthFactor;
-        }
-        return *this;
-    }
-
-    // Destructor
-    ~Fish()
-    {
-    }
-
-    // Metode
-    bool canEat(const Fish &other) const
-    {
-        return size > other.size;
-    }
-
-    void grow()
-    {
-        size = size + 1;
-        speed += 10;
-    }
-
-    void evolve()
-    {
-        if (size > 6)
-        {
-            name = "Reptila REMO";
-            speed += 20;
-            Logger::logEvent("Jucatorul a evoluat in Reptila!");
-        }
-        if (size > 11)
-        {
-            name = "Amfibianul.";
-            speed += 30;
-            Logger::logEvent("Jucatorul a evoluat in Amfibian hihi!");
-        }
-    }
-
-    const std::string &getName() const { return name; }
-
-    friend std::ostream &operator<<(std::ostream &os, const Fish &f)
-    {
-        os << "Fish(" << f.name << ", Size: " << f.size << ", Speed: " << f.speed << ")";
-        return os;
-    }
-
-    int getSize() const { return size; }
-    // int getSpeed() const { return speed; }
-};
-
-// Clasa Rewards
-class Reward
-{
-
-
-protected:
-    std::string type;
-    int value;
-    int turns;
-
-public:
-    Reward(){};
-    explicit Reward(const std::string &type, int value = 1, int turns = 1)
-        : type(type), value(value), turns(turns)
-    {
-    }
-
-    explicit Reward(const Reward &r)
-        : type(r.type), value(r.value), turns(r.turns)
-    {
-    }
-
-    Reward &operator=(const Reward &r)
-    {
-        if (this != &r)
-        {
-            type = r.type;
-            value = r.value;
-            turns = r.turns;
-        }
-        return *this;
-    }
-
-
-    virtual ~Reward() =  default;
-
-
-    friend std::ostream &operator<<(std::ostream &os, const Reward &r)
-    {
-        os << "Reward(" << r.type << ", Value: " << r.value << ", Turns: " << r.turns << ")" << std::endl;;
-        return os;
-    }
-    virtual bool isInvincible() const//daca o comentez typeid functioneaaza diferit ptr ca nu mai am polimorfism
-    {
-
-        return false;
-    }
-    virtual bool DoublePoints() const//daca o comentez typeid functioneaaza diferit ptr ca nu mai am polimorfism
-   {
-        return false;
-   }
-
-
-};
-
-class Invincible : public Reward
-{
-public:
-    Invincible(){};
-    using Reward::Reward;//inherits the constructors from rewards
-    bool isInvincible() const override
-    {
-        return true; //daca e invincibil poate sa manance ce peste vrea ptr o tura}
-    }
-
-
-};
-
-class isDoublePoints : public Reward
-{
-public:
-    isDoublePoints(){};
-    using Reward::Reward;
-    bool DoublePoints() const override
-    {
-        return true; //manaci un peste si primesti scor dublu
-    }
-
-
-};
 
 // Clasa Aquarium
-class Aquarium
-{
-    std::vector<Fish> fishies;
-    std::vector<std::shared_ptr <Reward>> rewards;
-    std::vector<int> size = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12};
 
-public:
-    Aquarium() = default; //ptr ca am alti constr
-
-
-    explicit Aquarium(const Aquarium &a)
-        : fishies(a.fishies), rewards(a.rewards)
-    {
-    }
-
-    Aquarium &operator=(const Aquarium &other)
-    {
-        if (this != &other)
-        {
-            fishies = other.fishies;
-            rewards = other.rewards;
-        }
-        return *this;
-    }
-
-
-    ~Aquarium()
-    {
-    }
-
-
-    void addFish(const Fish &fish) { fishies.push_back(fish); }
-    const std::vector<Fish> &getFishies() const { return fishies; }
-
-    void removeFish(int index)
-    {
-        if (index >= 0 && index < static_cast<int>(fishies.size()))
-        {
-            fishies.erase(fishies.begin() + index);
-        }
-    }
-
-    friend std::ostream &operator<<(std::ostream &os, const Aquarium &aquarium)
-    {
-        os << "Acvariul contine urmatorii pesti:\n";
-        const auto &fishList = aquarium.getFishies(); // Accesam lista de pesti
-        if (fishList.empty())
-        {
-            os << "Nu exista pesti in acvariu.\n";
-        } else
-        {
-            int cnt = 0;
-            for (const auto &fish: fishList)
-            {
-                os << cnt << '.' << fish << std::endl;
-                // Afișăm fiecare pește folosind operatorul << definit pentru Fish
-                ++cnt;
-            }
-        }
-        return os;
-    }
-
-
-    void addReward( std::shared_ptr<Reward> &r)
-    {
-        rewards.push_back(r);
-    }
-
-     std::vector<std::shared_ptr<Reward>> &getRewards()  { return rewards; }
-
-    void removeReward(int index)
-    {
-        if (index >= 0 && index < static_cast<int>(rewards.size()))
-        {
-            rewards.erase(rewards.begin() + index);
-        }
-    }
-};
 
 // Clasa Objective
 class Objective
